% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multiGene}
\alias{multiGene}
\title{multiGene}
\format{
data.frame
}
\usage{
multiGene
}
\value{
data.frame
}
\description{
A data.frame where the row names represent genes and each 
column represents the p-value of a particular omics analysis.
}
\keyword{datasets}
