% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNetEnrich.R
\name{multiNetEnrich}
\alias{multiNetEnrich}
\title{multiNetEnrich}
\usage{
multiNetEnrich(
  multiGene,
  network,
  p = 0,
  TERM2GENE = NULL,
  TERM2NAME = NULL,
  threshold = 1e-09,
  n = 10,
  nperm = 100,
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  qvalueCutoff = 0.2,
  combineMethod = "fisher",
  stoufferWeights = NULL
)
}
\arguments{
\item{network}{network}

\item{p}{restart probability}

\item{TERM2GENE}{TERM2GENE}

\item{TERM2NAME}{TERM2NAME}

\item{threshold}{threshold}

\item{n}{n}

\item{nperm}{Number of permutations to do.}

\item{combineMethod}{The method of combining pvalues.}

\item{genelist}{genelist}
}
\description{
multiNetEnrich
}
