% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TERM2GENE}
\alias{TERM2GENE}
\title{TERM2GENE}
\format{
data.frame
}
\usage{
TERM2GENE
}
\value{
data.frame
}
\description{
A data.frame of two column, the first column are terms, and the second column are genes.
}
\keyword{datasets}
